+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Smartness.java
@@ -1,5 +1,6 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.insanelib.util.MathHelper;
 import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
@@ -9,7 +10,10 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.LootBonusEnchantment;
+import net.minecraftforge.event.entity.living.LivingExperienceDropEvent;
+import net.minecraftforge.event.level.BlockEvent;
 
 public class Smartness extends Enchantment implements IEnchantmentTooltip {
     public Smartness() {
@@ -38,6 +42,20 @@ public class Smartness extends Enchantment implements IEnchantmentTooltip {
 
     public static int getIncreasedExperience(RandomSource random, int lvl, int experience) {
         return MathHelper.getAmountWithDecimalChance(random, experience * (1f + getBonusExperience(lvl)));
+    }
+
+    public static void applyToLivingDrops(LivingExperienceDropEvent event) {
+        if (event.getAttackingPlayer() == null)
+            return;
+        int lvl = EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.SMARTNESS.get(), event.getAttackingPlayer());
+        if (lvl > 0)
+            event.setDroppedExperience(Smartness.getIncreasedExperience(event.getEntity().getRandom(), lvl, event.getDroppedExperience()));
+    }
+
+    public static void applyToBlockDrops(BlockEvent.BreakEvent event) {
+        int lvl = EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.SMARTNESS.get(), event.getPlayer());
+        if (lvl > 0)
+            event.setExpToDrop(Smartness.getIncreasedExperience(event.getLevel().getRandom(), lvl, event.getExpToDrop()));
     }
 
     @Override